% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\docType{data}
\name{cats}
\alias{cats}
\title{my cats date}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{http://....org}{Link}
}
\usage{
data(cats)
}
\description{
Longer description of my data
}
\examples{
data("cats")
}
\references{
Put reference here
}
\keyword{datasets}
